Move ENABLE_RASTER_REPROJECTION to a temporary place
