+++ src/ol/reproj/common.js
@@ -3,3 +3,10 @@
  *     reprojection triangulation.
  */
 export var ERROR_THRESHOLD = 0.5;
+
+/**
+ * TODO: decide if we want to expose this as a build flag or remove it
+ * @type {boolean} Enable automatic reprojection of raster sources. Default is
+ *     `true`.
+ */
+export var ENABLE_RASTER_REPROJECTION = true;
